backend/requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Werkzeug==2.3.7
openpyxl==3.1.2
requests==2.31.0
